requirements.txt
torch>=2.1.0+cu128
transformers>=4.40.0
accelerate>=0.27.0
peft>=0.10.0                 
trl>=0.8.6                     
datasets>=2.18.0               
 
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
huggingface_hub>=0.22.0        
scikit-learn>=1.3.0
scipy>=1.11.04

llama-cpp-python==0.3.16